requirements.txt
flask==2.3.3
numpy==1.24.3
pillow==10.0.0
tensorflow==2.13.0
scikit-learn==1.3.0
opencv-python==4.8.0.76
matplotlib==3.7.2
pandas==2.0.3
gunicorn==21.2.0
flask-wtf==1.1.1